# packages/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Basalt API",
    "version": "1.1.0",
    "description": "Pro-tier hosted backend for Basalt. Runs on Cloudflare Workers; Open-tier surfaces (CLI, plugin, MCP, desktop) do not call this API.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/plsft/basalt/blob/main/LICENSE"
    },
    "contact": {
      "name": "Basalt",
      "url": "https://basalt.dev"
    }
  },
  "servers": [
    {
      "url": "https://api.basalt.dev",
      "description": "Production"
    },
    {
      "url": "https://api-staging.basalt.dev",
      "description": "Staging"
    }
  ],
  "components": {
    "schemas": {
...
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "Vault": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sync_enabled": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "user_id",
          "name",
          "created_at",
          "updated_at"
        ]
      },
      "VaultList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Vault"
        }
      },
      "VaultCreateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "enum": [
              "upload",
              "sync"
            ]
          }
        },
        "required": [
          "name"
        ]
      },
...
      "VaultSnapshot": {
        "type": "object",
        "description": "Client-built snapshot of the user's local index — notes, embeddings (base64-float32), and wikilinks.",
        "properties": {
          "schema": {
            "type": "integer",
            "const": 1
          },
          "vault_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "today": {
            "type": "string",
            "description": "ISO date the engine should treat as today."
          },
          "notes": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "rel_path": {
                  "type": "string"
                },
                "stem": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "created": {
                  "type": "string",
                  "nullable": true
                },
                "updated": {
                  "type": "string",
                  "nullable": true
                },
                "word_count": {
                  "type": "integer",
                  "minimum": 0
                },
                "content": {
                  "type": "string"
                },
                "content_hash": {
                  "type": "string"
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "rel_path",
                "stem",
                "title",
                "word_count",
                "content",
                "content_hash",
                "tags"
              ]
            }
          },
          "embeddings": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "rel_path": {
                  "type": "string"
                },
                "model": {
                  "type": "string"
                },
                "dim": {
                  "type": "integer",
                  "minimum": 1
                },
                "vec_b64": {
                  "type": "string",
                  "description": "Base64-encoded little-endian float32."
                }
              },
              "required": [
                "rel_path",
                "model",
                "dim",
                "vec_b64"
              ]
            }
          },
          "links": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "from_rel_path": {
                  "type": "string"
                },
                "target": {
                  "type": "string"
                }
              },
              "required": [
                "from_rel_path",
                "target"
              ]
            }
          }
        },
        "required": [
          "schema",
          "vault_id",
          "created_at",
          "notes",
          "embeddings",
          "links"
        ]
      },
      "VaultSnapshotAck": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "vault_id": {
            "type": "string"
          },
          "note_count": {
            "type": "integer"
          },
          "embedding_count": {
            "type": "integer"
          },
          "link_count": {
            "type": "integer"
          },
          "bytes": {
            "type": "integer"
          }
        },
        "required": [
          "ok",
          "vault_id",
          "note_count",
          "embedding_count",
          "link_count",
          "bytes"
        ]
      },
...
      "ReindexResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "vault_id": {
            "type": "string"
          },
          "note_count": {
            "type": "integer"
          },
          "vectors_upserted": {
            "type": "integer"
          },
          "embedding_model": {
            "type": "string"
          },
          "elapsed_ms": {
            "type": "integer"
          }
        },
        "required": [
          "ok",
          "vault_id",
          "note_count",
          "vectors_upserted",
          "embedding_model",
          "elapsed_ms"
        ]
      },
...
      "SnapshotMeta": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time"
          },
          "size": {
            "type": "integer"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "key",
          "size"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      }
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "basalt_session",
        "description": "HttpOnly session cookie set by /v1/auth/oauth/callback."
      }
    }
  },
  "paths": {
...
    "/v1/vaults": {
      "get": {
        "summary": "List the user's vaults",
        "description": "Excludes soft-deleted vaults.",
        "tags": [
          "Vaults"
        ],
        "responses": {
          "200": {
            "description": "Vault list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultList"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "post": {
        "summary": "Register a new vault",
        "description": "Creates a vault row and enqueues an indexing job.",
        "tags": [
          "Vaults"
        ],
        "responses": {
          "201": {
            "description": "Vault created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vault"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VaultCreateRequest"
              }
            }
          }
        }
      }
    },
    "/v1/vaults/:id": {
      "delete": {
        "summary": "Soft-delete a vault",
        "description": "30-day grace before hard purge. Findings + briefs are retained until the grace window expires.",
        "tags": [
          "Vaults"
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/v1/vaults/:id/snapshot": {
      "post": {
        "summary": "Upload a vault snapshot",
        "description": "Stores a client-built VaultSnapshot (notes + embeddings + links) in R2 keyed by user_id/vault_id. /v1/briefs/generate uses the latest snapshot.",
        "tags": [
          "Vaults"
        ],
        "responses": {
          "200": {
            "description": "Stored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultSnapshotAck"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VaultSnapshot"
              }
            }
          }
        }
      }
    },
    "/v1/vaults/:id/snapshot/meta": {
      "get": {
        "summary": "Snapshot metadata",
        "description": "Returns size, upload time, and custom metadata for the latest snapshot.",
        "tags": [
          "Vaults"
        ],
        "responses": {
          "200": {
            "description": "Snapshot meta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotMeta"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/v1/vaults/:id/reindex": {
      "post": {
        "summary": "Re-embed notes into Vectorize for cross-vault search",
        "description": "Reads the latest snapshot from R2, re-embeds title+content via Workers AI under the canonical model, and upserts the vectors into VECTORIZE with user/vault metadata.",
        "tags": [
          "Vaults"
        ],
        "responses": {
          "200": {
            "description": "Reindexed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReindexResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  }
}